createSnapshot: fix 64bit support
Finding tokens would not work correctly due to a 64bit value being
stored in a 32bit variable.

+++ cpctools/lib/CCPCSnapshot.cpp
@@ -301,7 +301,7 @@ void CCPCSnapshot::saveSnapshotIni(const string &filename) const
 CCPCSnapshot::TTokenID CCPCSnapshot::GetToken(const char *i_dataID, int &o_index)
 {
 	std::string token(i_dataID);
-	unsigned int p;
+	size_t p;
 	o_index = -1;
 	if ((p=token.find(':')) != std::string::npos)
 	{
